Fix individual client rate limiting with tc
The upload and download rates for rate limiting are determined via:

  authenticate_client()->do_binauth()

And they are then later stored in the client object at the end of
authenticate_client(). However, when using tc the according tc rules
with the per client limits are already set in between these two cases,
via:

  authenticate_client()
   -> auth_client_auth()
     -> auth_client_auth_nolock()
       -> auth_change_state()
         -> iptables_fw_authenticate()
           -> tc_attach_client()

Which will lead to the global rate limit being configured instead of the
individual per client one, deterimined via binauth, which is not yet
available in the client object.

Fixing this by moving the client authentication state machine change via
auth_client_auth() to after the new rate limits have been stored for the
client.

Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>