Fix quota + rate limits when switching to new account where disabled
When a user first logs in to an account which has a rate limit or quota
and then logs in with a different account which does not have a rate
limit or quota then this setting will wrongly not be updated.

Instead always update these client settings with every new session.

Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>